+++ include/Utility.hpp
@@ -8,13 +8,13 @@ namespace util
 {
 namespace detail
 {
-glm::vec2 pointToOpenGL(glm::ivec2 point, float half_width, float half_height);
-glm::vec2 vectorToOpenGL(glm::ivec2 vector, float half_width, float half_height);
+glm::vec2 pointToOpenGL(glm::vec2 point, float half_width, float half_height);
+glm::vec2 vectorToOpenGL(glm::vec2 vector, float half_width, float half_height);
 } // namespace detail
 
 void initialize(unsigned width, unsigned height);
 
-inline std::function<glm::vec2(glm::ivec2)> pointToOpenGL = nullptr;
-inline std::function<glm::vec2(glm::ivec2)> vectorToOpenGL = nullptr;
+inline std::function<glm::vec2(glm::vec2)> pointToOpenGL = nullptr;
+inline std::function<glm::vec2(glm::vec2)> vectorToOpenGL = nullptr;
 
 } // namespace util
